+++ source/frontends/libretro/game.cpp
@@ -78,7 +78,7 @@ namespace ra2
       const bool bVideoUpdate = true;
       const UINT dwClksPerFrame = NTSC_GetCyclesPerFrame();
 
-      const uint64_t cyclesToExecute = mySpeed.getCyclesTillNext(ourFrameTime);
+      const DWORD cyclesToExecute = mySpeed.getCyclesTillNext(ourFrameTime);
       const DWORD executedCycles = CpuExecute(cyclesToExecute, bVideoUpdate);
 
       g_dwCyclesThisFrame = (g_dwCyclesThisFrame + executedCycles) % dwClksPerFrame;
